Move calcGreatCircleDistance to the module
